requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
jinja2==3.1.4
python-multipart==0.0.12
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
asyncpg==0.29.0
databases[asyncpg]==0.9.0
sqlalchemy==2.0.36
openai==1.51.0
python-telegram-bot==21.6
authlib==1.3.2
httpx==0.27.2
python-dotenv==1.0.1
aiofiles==24.1.0
reportlab==4.2.5
apscheduler==3.10.4
itsdangerous==2.2.0
email-validator==2.2.0
Pillow==10.4.0
pydantic-settings==2.5.2
